requirements.txt
python-dotenv
fire
kafka-python
streamlit
datasets
transformers
torch
scikit-learn
numpy
pandas

